html/dom/documents/resource-metadata-management/document-lastModified.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>document.lastModified</title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel=help href="http://www.w3.org/html/wg/drafts/html/master/#resource-metadata-management">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  test(function(){
    var regexp = /^(0[1-9]|1[012])\/(0[1-9]|[12][0-9]|3[01])\/\d\d\d\d ([01][0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9])$/;
    assert_true(regexp.test(document.lastModified));
  }, "lastModified format");
</script>
